Simplify Error handling.
